[240215] - 스킬 밑작업

+++ Assets/Scripts/Skill.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 public class Skill : MonoBehaviour
 {
     public SkillType skillType;
-    public int value;
+    public float value;
     public float coolDown;
     public string line;
 }
